Power bi tool built for Associate Allocation
I have facing various situations in my life where I have been Thinking Big to achieve some crucial steps in my life .
Situation. One of the major situation has been while I started working for Receive Department as a Step Up Temp Lead. At this time I got an opportunity to reflect my skills and competencies in the field of technology and management and received a project from my Manager to Build a tool to simplify the operations happening in the various department in the FC.
Task. My task was to understand the process executed by the operations team with in the FC and propose the best fit solution for the same.
Action. In past as I have been experienced in Building digital tools, and Data visualization. I immediately acted over the same and build a prototype with in 2 weeks time on Power BI which can In take to manage all the Associate working on all the line, within receive, and showing there real time locations, productivity, cross trainings within the one single system. This project was no just made of one single department but can be replicated to different ,departments , direct and indirect process. In an estimate , it was saving 400 operational working hours per day within the FC, as it can be used by all the Leadership across all shifts and can save £8000 per day of 24 hours. By indirectly reducing the hustle of Leadership to manage the department on the analog systems of using associate batches on white board with stickers for there cross trainings .